// src/pages/DevicesCenter/data.d.ts
export type DeviceType = {
  activateTime: string;
  clientBuildNum: number;
  clientVersion: string;
  deviceName: string;
  deviceNumber: number;
  deviceType: number;
  dualStatus: any;
  effectivePolicyCount: number;
  effectiveQQWeChatCount: number;
  effectiveSMSCount: number;
  effectiveWebCount: number;
  eliminateTime: string;
  enablePerm: any;
  groupId: string;
  groupName: string;
  id: string;
  imei: string;
  imsi: string;
  knoxstatus: any;
  lastOnlineTime: string;
  loginId: string;
  mdmStatus: number;
  model: string;
  openUserId: string;
  os: string;
  policyStatus: any;
  productname: any;
  provisionFailCount: any;
  regionName: string;
  registTime: string;
  relationship: any;
  remark: any;
  rootFlag: number;
  scanStatus: any;
  sdSerialNum: any;
  serialNum: string;
  specialTypeFlag: any;
  status: number;
  totalPolicyCount: any;
  totalQQWeChatCount: any;
  totalSMSCount: any;
  totalWebCount: any;
  type: number;
  udid: string;
  userId: number;
  userName: string;
  versionNum: string;
};

export type DeviceDetailsType = {
  os: string;
  versionNum: string;
  onlineFlag: number;
  enablePermissionGuide: number;
  violationCauses: string;
  user: {
    groupId;
    groupName;
    id;
    loginId;
    mail;
    mobile;
    userName;
  };
  activateTime: string;
  udid: string;
  idfa: string;
  manufacturer: string;
  ram: string;
  wifiMac: string;
  model: string;
  romCapacity;
  powerStatus;
  cpu;
  serialNum;
  buildNumber;
};

export type SortType = {
  columnKey?: string;
  order?: string;
};